Update user command

+++ src/Bot/Telegram/Traits/Command.php
@@ -11,6 +11,17 @@ use App\MyAnimeList\MyAnimeList;
 
 trait Command
 {   
+    private function _user($args)
+    {
+        $args = trim($args);
+        $st = DB::pdoInstance()->prepare("SELECT * FROM `gm_user_warning` WHERE `uifd`=:uifd LIMIT 1;");
+        $st->execute([
+                ":uifd" => $this->input['message']['reply_to_message']['from']['id']."|".$this->room_id
+            ]);
+        $st = $st->fetch(PDO::FETCH_ASSOC);
+        B::sendMessage(json_encode($st, 128), $this->room_id, $this->msg_id);
+    }
+
     private function _nowarn($args)
     {
         $args = trim($args);
@@ -19,13 +30,15 @@ trait Command
                 ":uifd" => $this->input['message']['reply_to_message']['from']['id']."|".$this->room_id
             ]);
         $st = $st->fetch(PDO::FETCH_NUM);*/
-        if (isset($this->input['message']['reply_to_message']['from']['username'])) {
-            $user = "<a href=\"https://telegram.me/".$this->input['message']['reply_to_message']['from']['username']."\">".htmlspecialchars($this->input['message']['reply_to_message']['from']['first_name'])."</a>";
-        } else {
-            $user = "<code>".htmlspecialchars($this->input['message']['reply_to_message']['from']['first_name'])."</code>";
+        if (isset($this->input['message']['reply_to_message'])) {
+            if (isset($this->input['message']['reply_to_message']['from']['username'])) {
+                $user = "<a href=\"https://telegram.me/".$this->input['message']['reply_to_message']['from']['username']."\">".htmlspecialchars($this->input['message']['reply_to_message']['from']['first_name'])."</a>";
+            } else {
+                $user = "<code>".htmlspecialchars($this->input['message']['reply_to_message']['from']['first_name'])."</code>";
+            }
+            B::sendMessage("Done! {$user} has been forgiven", $this->room_id, $this->msg_id, ['parse_mode' => 'HTML', 'disable_web_page_preview'=>true]);
+            DB::pdoInstance()->prepare("DELETE FROM `gm_user_warning` WHERE `uifd`=:uifd LIMIT 1;")->execute([':uifd' => $this->input['message']['reply_to_message']['from']['id']."|".$this->room_id]);
         }
-        B::sendMessage("Done! {$user} has been forgiven", $this->room_id, $this->msg_id, ['parse_mode' => 'HTML', 'disable_web_page_preview'=>true]);
-        DB::pdoInstance()->prepare("DELETE FROM `gm_user_warning` WHERE `uifd`=:uifd LIMIT 1;")->execute([':uifd' => $this->input['message']['reply_to_message']['from']['id']."|".$this->room_id]);
     }
 
     private function _report($args)
